package/virtual: prevent deletes outside the virtual filesystem root (#337)
* package/virtual: prevent deletes outside the virtual filesystem root

* package/vfs: add deprecation note and uncomment test

+++ package/virtual/os.go
@@ -16,23 +16,35 @@ var _ FS = (OS)("")
 
 func (dir OS) Open(name string) (fs.File, error) {
 	if !fs.ValidPath(name) {
-		return nil, &fs.PathError{Op: "open", Path: name, Err: fs.ErrInvalid}
+		return nil, &fs.PathError{Op: "Open", Path: name, Err: fs.ErrInvalid}
 	}
 	return os.Open(filepath.Join(string(dir), name))
 }
 
 func (dir OS) MkdirAll(path string, perm fs.FileMode) error {
+	if !fs.ValidPath(path) {
+		return &fs.PathError{Op: "mkdirall", Path: path, Err: fs.ErrInvalid}
+	}
 	return os.MkdirAll(filepath.Join(string(dir), path), perm)
 }
 
 func (dir OS) WriteFile(name string, data []byte, perm fs.FileMode) error {
+	if !fs.ValidPath(name) {
+		return &fs.PathError{Op: "WriteFile", Path: name, Err: fs.ErrInvalid}
+	}
 	return os.WriteFile(filepath.Join(string(dir), name), data, perm)
 }
 
 func (dir OS) RemoveAll(path string) error {
+	if !fs.ValidPath(path) {
+		return &fs.PathError{Op: "RemoveAll", Path: path, Err: fs.ErrInvalid}
+	}
 	return os.RemoveAll(filepath.Join(string(dir), path))
 }
 
 func (dir OS) Sub(subdir string) (FS, error) {
+	if !fs.ValidPath(subdir) {
+		return nil, &fs.PathError{Op: "Sub", Path: subdir, Err: fs.ErrInvalid}
+	}
 	return OS(filepath.Join(string(dir), subdir)), nil
 }
